Refactor DashboardLayout and AppSidebar for improved structure and navigation
- Replace AppSidebar with AppSidebarWrapper in DashboardLayout to streamline layout management.
- Remove unused state and effects related to messaging and calls in DashboardLayout for cleaner code.
- Update AppSidebar to use Next.js Link component for navigation, enhancing client-side routing.
- Improve dropdown menu functionality in NavProjects for better user interaction and accessibility.
- Refactor ProfileCompletionCheck to enhance loading state management and user experience.
