+++ src/components/app-sidebar.tsx
@@ -17,6 +17,7 @@ import {
   Users
 } from "lucide-react"
 import { useSession } from "next-auth/react"
+import Link from "next/link"
 import { usePathname, useRouter } from "next/navigation"
 import * as React from "react"
 
@@ -106,7 +107,7 @@ export function AppSidebar({ showMessages = false, setShowMessages, ...props }: 
         if (setShowMessages) setShowMessages(false)
       }
     },
-  ], [pathname, showMessages, setShowMessages])
+  ], [pathname, showMessages, setShowMessages, toggleMessages])
 
   const getMenteeNavItems = React.useMemo(() => [
     {
@@ -214,7 +215,7 @@ export function AppSidebar({ showMessages = false, setShowMessages, ...props }: 
         <SidebarMenu>
           <SidebarMenuItem>
             <SidebarMenuButton size="lg" asChild>
-              <a href="/">
+              <Link href="/">
                 <div className="flex aspect-square size-8 items-center justify-center rounded-lg bg-indigo-600 text-white">
                   <Command className="size-4" />
                 </div>
@@ -222,7 +223,7 @@ export function AppSidebar({ showMessages = false, setShowMessages, ...props }: 
                   <span className="truncate font-semibold">MentorConnect</span>
                   <span className="truncate text-xs">{session?.user?.role || "User"}</span>
                 </div>
-              </a>
+              </Link>
             </SidebarMenuButton>
           </SidebarMenuItem>
         </SidebarMenu>
