+++ src/components/nav-main.tsx
@@ -47,12 +47,12 @@ export function NavMain({
             <SidebarMenuItem>
               <SidebarMenuButton asChild tooltip={item.title}>
                 {item.onClick ? (
-                  <Link href={item.url} onClick={item.onClick} className="flex h-9 w-full items-center gap-2 rounded-md px-3 hover:bg-sidebar-hover">
+                  <Link href={item.url} onClick={item.onClick} scroll={false}>
                     <item.icon />
                     <span>{item.title}</span>
                   </Link>
                 ) : (
-                  <Link href={item.url} className="flex h-9 w-full items-center gap-2 rounded-md px-3 hover:bg-sidebar-hover">
+                  <Link href={item.url} scroll={false}>
                     <item.icon />
                     <span>{item.title}</span>
                   </Link>
@@ -72,11 +72,11 @@ export function NavMain({
                         <SidebarMenuSubItem key={subItem.title}>
                           <SidebarMenuSubButton asChild>
                             {subItem.onClick ? (
-                              <Link href={subItem.url} onClick={subItem.onClick} className="flex h-9 w-full items-center gap-2 rounded-md px-3 pl-10 hover:bg-sidebar-hover">
+                              <Link href={subItem.url} onClick={subItem.onClick} scroll={false}>
                                 <span>{subItem.title}</span>
                               </Link>
                             ) : (
-                              <Link href={subItem.url} className="flex h-9 w-full items-center gap-2 rounded-md px-3 pl-10 hover:bg-sidebar-hover">
+                              <Link href={subItem.url} scroll={false}>
                                 <span>{subItem.title}</span>
                               </Link>
                             )}
